Reword the entire front page to be more stern.
Of the 669 currently pending domains, 75 of them no longer satisfy the requirements that they satisfied at submission time. This is for a variety of reasons:

11 domain.tls.cannot_connect
13 domain.tls.invalid_cert_chain
7  domain.www.no_tls
7  header.preloadable.include_sub_domains.missing
6  header.preloadable.preload.missing
2  internal.redirects.http.first_probe_failed
4  redirects.follow_error
1  redirects.http.first_redirect.insecure
1  redirects.http.first_redirect.invalid
9  redirects.http.no_redirect
9  redirects.http.www_first
2  redirects.insecure.subsequent
4  redirects.too_many
1  response.multiple_headers
1  response.no_header

It's hard to tell if these are by accident, or because they don't care, or because they think they can get away with loosening their TLS setup after submission. All of these possible reasons are disappointing, because the HSTS preload list is growing quickly, and we don't want to waste space on sites that aren't actually committed to HTTPS.

Until we figure out our policy for removal, the text changes in this commit are meant to make our expectations crystal-clear.