Increase database timeout and add logging to /api/v2/update
This is a stop-gap measure to help with issue #154 until a more scalable
solution is designed. The logging in /api/v2/update times how long it
takes to fetch the preload list from Chromium and how long it takes to
load preloaded domains from the database. The latter currently takes
around 15 seconds for approximately 50,000 entries.