Include the domain on the preload list in the status API call.
The status API call will return the effective status of a domain, which
includes applying the status of a parent domain with includeSubDomains
set if there is no entry for the queried domain. This change adds a
"preloadedDomain" field to the status response if the domain is
preloaded, which indicates the domain name on the preload list that
resulted in indicating the domain is preloaded.