Created new domain status: StatusPendingAutomatedRemoval (#226)
With this new domain status, we will be able to properly check for
domains that have failed the preload list checks too many times in our
automated removal process, and thus should be removed from the preload
list. We change these domains' status to StatusPendingAutomatedRemoval.
We will use this status in the python script that generates a new source
preload list by ensuring that domains that are marked pendingRemoval in
the script includes domains that have requested removal from the preload
list and the domains that have failed the automated checks too many
times.

This PR doesn't include any logic concerning when the domains will have
their status' changed to StatusPendingAutomatedRemoval. That part is
being created by my partner, @bkhushi.