Write new tests for Update API endpoint (#259)
Switch to table-driven tests, make it more clear what each test is
testing, and add additional test coverage. All tests pass to document
the current behavior of the Update endpoint. Some of the behavior of
Update is incorrect; tests for those are commented with TODOs indicating
what their correct behavior should be.