Reword content on hstspreload.org
Many web browsers will upgrade HTTP navigations to HTTPS by default now,
which lessens the value proposition of preloading HSTS (vs HSTS without
preloading). Because of this, we don't recommend that sites preload
their domains.

Part of this change also modifies the styling of the submission form to
have elements generally sized closer to body text size and keep text
left-justified instead of centered, to match the rest of the page.