Update dependencies
go.mod and go.sum were updated by running `go get -u && go mod tidy`.
Since the update to go.mod included an update to the go version, this
change also updates the CI config to the same version.

The update to dependencies results in deprecating a function we're
calling, so this change also adds a comment to disable the lint check
on that deprecated function call until it can be fixed properly.