Create new tool to manage updating the list in chromium src
The new scripts/updatelist tool is an intended replacement for
scripts/roll_preload_list.py. This updatelist tool also incorporates
the functionality of the `hstspreload scan-pending` command that has to
be run as a prerequisite to roll_preload_list.py. It additionally
performs revalidation checks on the domains from the
/api/v2/pending-automated-removal list.